adjust GDB/Insight tarball name of version 6.0-7.3
A mistake has been detected in the release tar files for all
GDB releases from version 6.0 to version 7.3 (included). The
mistake has been corrected, and the FSF issued the following
announcements:
  - http://www.sourceware.org/ml/gdb/2011-09/msg00135.html
  - http://www.sourceware.org/ml/gdb/2011-09/msg00136.html

As an result of this all tarballs were renamed. Each package
and only the tarball package (!) has an extended version number
string with a small 'a' on end.

Signed-off-by: Stephan Linz <linz@li-pro.net>