libc/uClibc: add multilib-cacpable backend
For mutlilib, the C library must be built once for each variants.

When doing multilib, we only need the headers from the default
variant, but we need the startfiles for each variants. Allow the
frontend to specify either one, or both.

Special care must be taken to put the resulting libraries in the
proper places. When building a multilib variant, install in a
separate directory, to avoid clutering the default or any other
variant.

Signed-off-by: Stephan Linz <linz@li-pro.net>